pyproject.toml
[project]
name = "qrcodekit"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "ajrlewis", email = "hello@ajrlewis.com" }
]
dependencies = [
    "pillow==10.2.0",
    "pypng==0.20220715.0",
    "qrcode==7.4.2",
    "typing_extensions==4.9.0"
]
license = "MIT"

[tool.setuptools]
packages = ["qrcodekit", "fonts", "images", "scripts"]
